# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para consultar alumnos de mi centro",
        "description": "Esta api permite interactuar con los alumnos de bd del instituto",
        "contact": {
            "name": "Tabata",
            "email": "tabata.lopez11@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/mit"
        },
        "version": "2.0.0"
    },
    "paths": {
        "/api/alumnos": {
            "get": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Obtener todos los alumnos",
                "operationId": "GetAlumnos",
                "responses": {
                    "200": {
                        "description": "Éxito",
                        "content": {
                            "application/vnd.api+json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Alumno"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Actualizar datos de alumno",
                "description": "Updates user data in the database",
                "operationId": "updateAlumno",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Crear un nuevo alumno",
                "description": "Stores a new user in the database",
                "operationId": "createAlumno",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "johndoe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Alumnos"
                ],
                "summary": "Eliminar un alumno",
                "description": "Deletes a user in the database",
                "operationId": "deleteAlumno",
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Alumno": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "example": "juan@example.com"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Alumnos",
            "description": "Alumnos"
        }
    ]
}